refactor(core): move task.builder into task folder
